Persist form field values on error.
When a form is submitted and it has validation errors, the resulting
page now keeps the values that the user has entered so that they only
have to fix the fields with errors.

This fix is made easier by taking advantage of WTForms provided HTML
rendering which eliminates a lot of the HTML we had provided ourselves.

Fixes #20